+++ tests/conftest.py
@@ -5,7 +5,7 @@ import shutil
 import sys
 from contextlib import contextmanager
 from pathlib import Path
-from typing import Dict, Generator, Iterable
+from typing import Generator, Iterable
 
 import pytest
 from pkginfo import Wheel
@@ -31,7 +31,7 @@ def cd(p: Path) -> Generator[None, None, None]:
     os.chdir(old)
 
 
-def hashes(directory: Path) -> Dict[Path, str]:
+def hashes(directory: Path) -> dict[Path, str]:
     hs = {}
     for file in directory.rglob("*"):
         if str(file).startswith("."):
@@ -121,7 +121,7 @@ def test_ep():
 [project]
 name = "testproject"
 description = "an example test project for testing vulcan builds, %"
-authors = [{{name="Joel Christiansen", email="joelchristiansen@optiver.com"}}]
+authors = [{name="Joel Christiansen", email="joelchristiansen@optiver.com"}]
 keywords = [ "build", "testing" ]
 classifiers = [
     "Topic :: Software Development :: Build Tools",
@@ -150,33 +150,18 @@ test = ['pytest']
 lint = ['flake']
 
 [tool.vulcan.dependencies]
-requests = {{version="~=2.25.1", extras=["security"]}}
+requests = {version="~=2.25.1", extras=["security"]}
 
 [tool.vulcan.extras]
 test1 = ["requests", "build", "packaging~=20.9"]
 test2 = ["requests~=2.22", "setuptools"]
 test3 = ["requests>=2.0.0", "wheel"]
 
-[[tool.vulcan.shiv]]
-bin_name="testproject"
-console_script="myep"
-interpreter='{cur_interp}'
-extra_args="-E --compile-pyc"
-
-[[tool.vulcan.shiv]]
-bin_name="testproject2"
-console_script="myep"
-interpreter='{cur_interp}'
-extra_args="-E --compile-pyc"
-
 [build-system]
 requires=['setuptools', 'vulcan-py']
 build-backend="vulcan.build_backend"
 
-""".format(
-                cur_interp=sys.executable
-            )
-        )
+""")
 
     return tmp_path
 
@@ -192,7 +177,7 @@ def test_application_forbidden_keys(tmp_path_factory: pytest.TempPathFactory) ->
     tmp_path = tmp_path_factory.mktemp("build_testproject")
     (tmp_path / "testproject").mkdir()
     (tmp_path / "testproject/__init__.py").write_text(
-        """\
+        """
 def test_ep():
     print("Running!")
 """
@@ -206,8 +191,8 @@ def test_ep():
             """\
 [project]
 name = "testproject"
-description = "an example test project for testing vulcan builds, %"
-authors = [{{name="Joel Christiansen", email="joelchristiansen@optiver.com"}}]
+description = "an example test project for testing vulcan builds"
+authors = [{name="Joel Christiansen", email="joelchristiansen@optiver.com"}]
 keywords = [ "build", "testing" ]
 classifiers = [
     "Topic :: Software Development :: Build Tools",
@@ -222,7 +207,6 @@ myep = "testproject:test_ep"
 foobar = "barfoo"
 module_dir = "testproject"
 
-
 [project.entry-points.test_eps]
 myep = "testproject:test_ep"
 
@@ -231,28 +215,17 @@ packages = [ "testproject" ]
 plugins = ['example_plugin']
 
 [tool.vulcan.dependencies]
-requests = {{version="~=2.25.1", extras=["security"]}}
+requests = {version="~=2.25.1", extras=["security"]}
 
 [tool.vulcan.extras]
 test1 = ["requests", "build"]
 test2 = ["requests~=2.22", "setuptools"]
 test3 = ["requests>=2.0.0", "wheel"]
 
-[[tool.vulcan.shiv]]
-bin_name="testproject"
-console_script="myep"
-interpreter='{cur_interp}'
-extra_args="-E --compile-pyc"
-
-
 [build-system]
 requires=['setuptools', 'vulcan-py']
 build-backend="vulcan.build_backend"
-
-""".format(
-                cur_interp=sys.executable
-            )
-        )
+""")
 
     return tmp_path