+++ vulcan/builder.py
@@ -1,15 +1,15 @@
 from __future__ import annotations
+
 import asyncio
 import tempfile
 from itertools import chain
-from typing import Dict, List, Tuple
 
 from pkg_resources import Requirement
 
 from vulcan.isolation import VulcanEnvBuilder, create_venv
 
 
-async def build_requires(pipenv: VulcanEnvBuilder, requires: List[str]) -> Dict[Requirement, Requirement]:
+async def build_requires(pipenv: VulcanEnvBuilder, requires: list[str]) -> dict[Requirement, Requirement]:
     with tempfile.TemporaryDirectory() as site_packages:
         await pipenv.install(site_packages, requires)
         freeze = await pipenv.freeze(site_packages)
@@ -17,10 +17,10 @@ async def build_requires(pipenv: VulcanEnvBuilder, requires: List[str]) -> Dict[
 
 
 async def resolve_deps(
-    install_requires: List[str],
-    extras: Dict[str, List[str]],
+    install_requires: list[str],
+    extras: dict[str, list[str]],
     python_version: str | None = None,
-) -> Tuple[List[str], Dict[str, List[str]]]:
+) -> tuple[list[str], dict[str, list[str]]]:
 
     if not install_requires and not extras:
         return [], {}